Refactors the ui into separate files/widgets
