+++ include/common.cpp
@@ -1,0 +1,6 @@
+#include "common.h"
+
+Common::Common()
+{
+
+}
