+++ include/function/normalizer.h
@@ -1,0 +1,34 @@
+#ifndef NORMALIZER_H
+#define NORMALIZER_H
+
+#include <QWidget>
+#include "common.h"
+
+class QListWidgetItem;
+
+namespace Ui {
+class Normalizer;
+}
+
+class Normalizer : public QWidget, public Common
+{
+    Q_OBJECT
+
+public:
+    explicit Normalizer(QWidget *parent = nullptr);
+    ~Normalizer();
+    QString output() const;
+signals:
+    void outputChanged();
+
+private slots:
+    void on_normalizer_generate_clicked();
+    void on_normalizer_bpmlist_itemClicked(QListWidgetItem *item);
+
+    void on_normalizer_output_textChanged();
+
+private:
+    Ui::Normalizer *ui;
+};
+
+#endif // NORMALIZER_H
